Tweak fz_color_params and fz_store_hash.
Use uint8_t for fz_color_param entries - no need to use
larger.

Same change in fz_store_hash. This limits the size required
for the hash table too.

+++ include/mupdf/fitz/colorspace.h
@@ -20,10 +20,10 @@ typedef struct fz_color_params_s fz_color_params;
 
 struct fz_color_params_s
 {
-	int ri;
-	int bp;
-	int op;
-	int opm;
+	uint8_t ri;
+	uint8_t bp;
+	uint8_t op;
+	uint8_t opm;
 };
